feat: создал функционал добавления колонок #304749 + fix: починил конвертацию order #305768 (#184)
* feat: создал кнопку добавления новой колонки

* refactor: переместил кнопку добавления колонки в отдельный файл

* refactor: переделал кнопку добавления стейджа в WstoreWidget

* добавил textField для ввода названия новой колонки

* feat: создал метод, который создает стейдж на беке

* feat: добавлен метод createStage в wstore

* fix: поправил конвертацию order

* fix: сделал локальное обновление колонок при добпалении

* refactor: убрал ненужный коммент

* refactor: сделал дополнительный метод, проверяющий на null id

+++ lib/models/task_models.dart
@@ -246,7 +246,7 @@ class TasksProjectGroup implements ITasksGroup {
   final int spaceOrder;
   @override
   final String groupTitle;
-  final int projectOrder;
+  final double projectOrder;
   final int columnOrder;
   final int spaceId;
   @override
@@ -296,7 +296,7 @@ class TasksDateGroup implements ITasksGroup {
 
 class TaskStageWithOrder {
   final String stageName;
-  final int stagesOrder;
+  final double stagesOrder;
   final int taskOrder;
 
   TaskStageWithOrder({
@@ -311,7 +311,7 @@ class TaskStageWithOrder {
 class SortedTask {
   final int id;
   final String stageName;
-  final int stageOrder;
+  final double stageOrder;
   final Task task;
   final int taskOrder;
